requirements.txt
pymongo==4.6.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
tensorflow==2.15.0
shap==0.44.1
lime==0.2.0.1
matplotlib==3.8.2
seaborn==0.13.0
fastapi==0.108.0
uvicorn==0.25.0
python-dotenv==1.0.0
joblib==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
schedule==1.2.1
apscheduler==3.10.4